refactor: v2 LoadingPage 적용
