+++ src/pages/generate/apis/mutations/useGenerateFullFunnelImageMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 
-import { useGenerateStore } from '@pages/generate/stores/useGenerateStore';
+import { useGenerateStore } from '@pages/generate/v2/stores/useGenerateStore';
 
 import type {
   GenerateImageV4Response,
