+++ src/pages/generate/v2/pages/loading/hooks/useGenerateImageRequest.ts
